Added batch training functionality instead of per-case training
- There is now NeuralNetwork::train_batch and NeuralNetwork::back_propagate_batch() for that, with added rolling proportion for training
- Removed the rolling proportion for classical training
- Added subsequent tests
- Made a main for testing on the whole set, with rolling proportion set to 25%

+++ tests/hexadecimal_parser.cpp
@@ -15,13 +15,13 @@
 TEST(BinaryWithSplit_Parser, Single_Instruction) {
 	std::string hexa = ";; Cycles: x\n\n"
 					   "FF FF\n";
-	std::vector<double> expected = {1.0, 1.0};
-	std::vector<double> result = HexadecimalParser().parse_in(hexa);
+	std::vector<float> expected = {1.0, 1.0};
+	std::vector<float> result = HexadecimalParser().parse_in(hexa);
 	EXPECT_EQ(result, expected);
 
 	hexa = ";; Cycles: x\n\n"
 		   "FF EE 11 0B\n";
-	expected = {1.0, (double)0xEE / 255.0, (double)0x11 / 255.0, (double)0x0B / 255.0};
+	expected = {1.0, (float)0xEE / 255.0, (float)0x11 / 255.0, (float)0x0B / 255.0};
 	result = HexadecimalParser().parse_in(hexa);
 	EXPECT_EQ(result, expected);
 }
@@ -36,15 +36,15 @@ TEST(BinaryWithSplit_Parser, Multiple_Instructions) {
 							   "00\n"
 							   "71 D9\n";
 
-	std::vector<double> expected;
-	std::vector<double> inst1 = {(double)0xAA / 255.0, (double)0xBB / 255.0};
-	std::vector<double> inst2 = {(double)0xC5 / 255.0, (double)0xE4 / 255.0, (double)0xB8 / 255.0};
-	std::vector<double> inst3 = {0.0};
-	std::vector<double> inst4 = {(double)0x71 / 255.0, (double)0xD9 / 255.0};
+	std::vector<float> expected;
+	std::vector<float> inst1 = {(float)0xAA / 255.0, (float)0xBB / 255.0};
+	std::vector<float> inst2 = {(float)0xC5 / 255.0, (float)0xE4 / 255.0, (float)0xB8 / 255.0};
+	std::vector<float> inst3 = {0.0};
+	std::vector<float> inst4 = {(float)0x71 / 255.0, (float)0xD9 / 255.0};
 	expected.insert(expected.end(), inst1.begin(), inst1.end());
 	expected.insert(expected.end(), inst2.begin(), inst2.end());
 	expected.insert(expected.end(), inst3.begin(), inst3.end());
 	expected.insert(expected.end(), inst4.begin(), inst4.end());
-	std::vector<double> result = HexadecimalParser().parse_in(instructions);
+	std::vector<float> result = HexadecimalParser().parse_in(instructions);
 	EXPECT_EQ(result, expected);
 }
